ES-Ingestion/requirements.txt
opensearch-py
aiohttp
python-dotenv